[SDK] Support validation for eager flow (#2061)
# Description

Please add an informative description that covers that changes made by
the pull request and link all relevant issues.
This pull request includes changes mainly to the
`src/promptflow/promptflow/_sdk/entities/_eager_flow.py` and
`src/promptflow/promptflow/_sdk/operations/_flow_operations.py` files,
with a focus on improving the validation and error handling of the
`EagerFlow` class. The most notable changes include the addition of the
`SchemaValidatableMixin` to the `EagerFlow` class, the modification of
the `_load` method to raise a validation error on unknown fields, and
the addition of a new test case for eager flow validation.

Changes to `EagerFlow` class:

*
[`src/promptflow/promptflow/_sdk/entities/_eager_flow.py`](diffhunk://#diff-53d7ff49a8720f52ee715a6f63d5e389750e44c578a907076d435e541f0163eeL6-R15):
The `EagerFlow` class now inherits from `SchemaValidatableMixin`,
allowing it to use its methods for schema validation. The `_load` method
has been modified to raise a validation error on unknown fields, and the
`_default_context`, `_create_validation_error`, and
`_dump_for_validation` methods have been added to support the new
functionality.
[[1]](diffhunk://#diff-53d7ff49a8720f52ee715a6f63d5e389750e44c578a907076d435e541f0163eeL6-R15)
[[2]](diffhunk://#diff-53d7ff49a8720f52ee715a6f63d5e389750e44c578a907076d435e541f0163eeR43-R77)

Changes to `FlowOperations` class:

*
[`src/promptflow/promptflow/_sdk/operations/_flow_operations.py`](diffhunk://#diff-afdd40a5d0519512dcf9be48bd46c4caaa2291b808687de77896989af63f47e4L665-R674):
The `validate` method now passes `raise_error=False` when loading the
flow entity. This change allows the method to continue execution even if
there are validation errors, which are then handled later in the method.
The method also now uses the `path` attribute of the flow entity instead
of `flow_dag_path` when resolving the location for diagnostics.
[[1]](diffhunk://#diff-afdd40a5d0519512dcf9be48bd46c4caaa2291b808687de77896989af63f47e4L665-R674)
[[2]](diffhunk://#diff-afdd40a5d0519512dcf9be48bd46c4caaa2291b808687de77896989af63f47e4L690-R692)

Changes to test cases:

*
[`src/promptflow/tests/sdk_cli_test/e2etests/test_flow_local_operations.py`](diffhunk://#diff-39113ef42bdaeb63710e4eaf72fec3120025601166fdfa766341820991ddf8a4R14-R22):
A new test case, `test_eager_flow_validate`, has been added to test the
validation of eager flows. This test case checks that the correct error
message is returned when the entry function of an eager flow is invalid.
[[1]](diffhunk://#diff-39113ef42bdaeb63710e4eaf72fec3120025601166fdfa766341820991ddf8a4R14-R22)
[[2]](diffhunk://#diff-39113ef42bdaeb63710e4eaf72fec3120025601166fdfa766341820991ddf8a4R482-R494)
# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.