[Executor] Add model for llm trace and handle token inputs for embedding trace (#2130)
# Description

Main Changes:
1. Add `llm.model` in the span attribute of llm trace;
2. For OpenAI embedding API, the inputs could be string or tokenized
string (tokenized string is a token array, which is a list of int), for
the string case, we update the original text in the span attribute:
```
"embedding.embeddings": [
    {
        "embedding.vector": <len(embedding) dimensional vector>",
        "embedding.text": original_text
    }
]
```
For the token array, we should update the length of token array in the
span attribute;
```
"embedding.embeddings": [
    {
        "embedding.vector": <len(embedding) dimensional vector>",
        "embedding.text": <len(token_array) dimensional token>
    }
]
```
5. Add tests for the new feature.


# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.

---------

Co-authored-by: Lina Tang <linatang@microsoft.com>