[SDK][azure]Add identity support for run for automatic runtime (#2157)
# Description

Please add an informative description that covers that changes made by
the pull request and link all relevant issues.
This pull request introduces identity support for automatic runtime in
the PromptFlow SDK and CLI. The changes include the addition of identity
configuration in the YAML schema, the introduction of new Enums for
identity configuration, and the addition of identity resolution in the
run operations.

Changes to YAML schema and documentation:

*
[`docs/reference/run-yaml-schema-reference.md`](diffhunk://#diff-f57019f0c08dc1936b08a86c6e3cd020676f614069559309a6b2216b7b9c89b3R29-R30):
Added a new `identity` field in the schema and updated the documentation
accordingly. The `identity` field is an object that supports two types
of identities: `managed` and `user_identity`.
[[1]](diffhunk://#diff-f57019f0c08dc1936b08a86c6e3cd020676f614069559309a6b2216b7b9c89b3R29-R30)
[[2]](diffhunk://#diff-f57019f0c08dc1936b08a86c6e3cd020676f614069559309a6b2216b7b9c89b3R39-R46)

*
[`src/promptflow/promptflow/_sdk/schemas/_run.py`](diffhunk://#diff-db38cc07d25efdbc0622e8e9352e07e34c502d36bdad1954bcd52db156192c9fR9-R14):
Introduced `ManagedIdentitySchema` and `UserIdentitySchema` for identity
configuration in the YAML schema. The `RunSchema` now includes an
`identity` field that can be either of the two new schema types.
[[1]](diffhunk://#diff-db38cc07d25efdbc0622e8e9352e07e34c502d36bdad1954bcd52db156192c9fR9-R14)
[[2]](diffhunk://#diff-db38cc07d25efdbc0622e8e9352e07e34c502d36bdad1954bcd52db156192c9fR39-R55)
[[3]](diffhunk://#diff-db38cc07d25efdbc0622e8e9352e07e34c502d36bdad1954bcd52db156192c9fR112-R118)

Changes to the SDK:

*
[`src/promptflow/promptflow/_sdk/_constants.py`](diffhunk://#diff-9ffef684cefef289f51e58b0aff1cd6cfa42363ace309d07f36dfe6559c5b84dR468-R485):
Added two new Enums `IdentityConfigurationType` and `IdentityKeys` for
identity configuration.

*
[`src/promptflow/promptflow/_sdk/entities/_run.py`](diffhunk://#diff-14f4d7a1d9b077c7b749230a92628782f990e1fc7cb741b20ff894f868327020R32):
The `Run` class now includes an `_identity` field. The `Run` class's
`__init__` and `_to_rest_object` methods have been updated to handle the
new identity configuration.
[[1]](diffhunk://#diff-14f4d7a1d9b077c7b749230a92628782f990e1fc7cb741b20ff894f868327020R32)
[[2]](diffhunk://#diff-14f4d7a1d9b077c7b749230a92628782f990e1fc7cb741b20ff894f868327020R185)
[[3]](diffhunk://#diff-14f4d7a1d9b077c7b749230a92628782f990e1fc7cb741b20ff894f868327020R557-R559)
[[4]](diffhunk://#diff-14f4d7a1d9b077c7b749230a92628782f990e1fc7cb741b20ff894f868327020R582)

*
[`src/promptflow/promptflow/azure/operations/_run_operations.py`](diffhunk://#diff-8edc3f1b002af9c7a0059d7d6ae7a25d1983836b6736d0d7e8c8537c6c3781c2R41):
The `RunOperations` class now includes a `_resolve_identity` method to
resolve the identity configuration for a run. The `__init__` method and
`_resolve_dependencies_in_parallel` method have been updated to handle
the new identity configuration.
[[1]](diffhunk://#diff-8edc3f1b002af9c7a0059d7d6ae7a25d1983836b6736d0d7e8c8537c6c3781c2R41)
[[2]](diffhunk://#diff-8edc3f1b002af9c7a0059d7d6ae7a25d1983836b6736d0d7e8c8537c6c3781c2R101)
[[3]](diffhunk://#diff-8edc3f1b002af9c7a0059d7d6ae7a25d1983836b6736d0d7e8c8537c6c3781c2R842)
[[4]](diffhunk://#diff-8edc3f1b002af9c7a0059d7d6ae7a25d1983836b6736d0d7e8c8537c6c3781c2R975-R1013)

Changes to the CLI:

*
[`src/promptflow/CHANGELOG.md`](diffhunk://#diff-41ec3f7c4b5d4c0e670407d3c00a03a6966d7ebf617b1536473e33a12e2bc765R13):
Updated the changelog to include the new identity support for automatic
runtime.

Changes to tests:

*
[`src/promptflow/tests/sdk_cli_azure_test/e2etests/test_run_operations.py`](diffhunk://#diff-b63fdccdd1a3a293abe7799c0a231a7226614f726b6de1298f7e0233af26f128R19-R20):
Added new tests to validate the identity configuration for automatic
runtime.
[[1]](diffhunk://#diff-b63fdccdd1a3a293abe7799c0a231a7226614f726b6de1298f7e0233af26f128R19-R20)
[[2]](diffhunk://#diff-b63fdccdd1a3a293abe7799c0a231a7226614f726b6de1298f7e0233af26f128R1144-R1193)

*
[`src/promptflow/tests/sdk_cli_azure_test/unittests/test_run_operations.py`](diffhunk://#diff-e61f9744c35acda81ccf7220c70aad98d0a258eec5be24997c151732a6d1ac0eR2-R15):
Added new unit tests to validate the identity resolution in
`RunOperations`.
[[1]](diffhunk://#diff-e61f9744c35acda81ccf7220c70aad98d0a258eec5be24997c151732a6d1ac0eR2-R15)
[[2]](diffhunk://#diff-e61f9744c35acda81ccf7220c70aad98d0a258eec5be24997c151732a6d1ac0eR26-R81)

Changes to test configurations:

*
[`src/promptflow/tests/test_configs/runs/sample_bulk_run_with_managed_identity.yaml`](diffhunk://#diff-62088854bccb57b83136d73c7396e8d65b23856d49886d992690669f5acb19e4R1-R10):
Added a new test configuration for a run with a managed identity.

*
[`src/promptflow/tests/test_configs/runs/sample_bulk_run_with_user_identity.yaml`](diffhunk://#diff-1d8af90c7da2c611161942ac3ed0d3acdda2b567ec76817ad1359251dabed0d2R1-R9):
Added a new test configuration for a run with a user identity.

# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.