doc: update doc for using flow in pipeline (#2164)
# Description

This pull request modifies the
`docs/cloud/azureai/use-flow-in-azure-ml-pipeline.md` file to provide a
more detailed and structured guide on using a flow as a parallel
component in a pipeline job on AzureML. The changes include updating the
pre-requirements, providing an example of a flow, explaining how to
register a flow as a component, and detailing the environment of the
component and its ports and run settings.

Main changes:

* Updated the pre-requirements section with a more structured guide on
enabling the feature. It now includes steps on installing the necessary
CLI or package, ensuring the `$schema` is in the target source, and
generating and updating the metadata.
* Provided an example of a flow and explained how to register it as a
component using either the CLI or SDK.
* Added a section on the environment of the component. It explains how
the environment is based on the latest promptflow runtime image and how
to use an existing Azure ML environment or define the environment in
Azure ML style.
* Added a section on component ports and run settings. It provides a
detailed table explaining the input ports, input parameters, output
ports, and run settings of the component.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [x] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [x] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.