[SDK] Support eager flow local deploy (#2186)
# Description

Since serving app depends on flow contract, added eager flow contract.
This pull request primarily introduces the concept of "EagerFlow" to the
PromptFlow SDK, which is a new type of flow that potentially improves
performance and simplifies the flow execution process. The changes also
include the refactoring of the `Flow` class into a `FlowBase` class and
two subclasses, `Flow` and `EagerFlow`. The `FlowInvoker` class is
updated to handle the initialization and execution of both types of
flows differently. Test cases are added to verify the behavior of the
new `EagerFlow`.

Here are the most important changes:

Refactoring of Flow classes:

*
[`src/promptflow/promptflow/contracts/flow.py`](diffhunk://#diff-b353941bce91518c7f494112d2dd5088b73681eb66e6ff19294edb9c3ec05d0fL520-R564):
The `Flow` class was refactored into a `FlowBase` class and two
subclasses, `Flow` and `EagerFlow`. The `FlowBase` class now contains
shared methods and properties of both subclasses.
[[1]](diffhunk://#diff-b353941bce91518c7f494112d2dd5088b73681eb66e6ff19294edb9c3ec05d0fL520-R564)
[[2]](diffhunk://#diff-b353941bce91518c7f494112d2dd5088b73681eb66e6ff19294edb9c3ec05d0fL543-L547)
[[3]](diffhunk://#diff-b353941bce91518c7f494112d2dd5088b73681eb66e6ff19294edb9c3ec05d0fL689-L699)
[[4]](diffhunk://#diff-b353941bce91518c7f494112d2dd5088b73681eb66e6ff19294edb9c3ec05d0fL814-R843)
[[5]](diffhunk://#diff-b353941bce91518c7f494112d2dd5088b73681eb66e6ff19294edb9c3ec05d0fL837-L843)
[[6]](diffhunk://#diff-b353941bce91518c7f494112d2dd5088b73681eb66e6ff19294edb9c3ec05d0fR887-R929)

Introduction of EagerFlow:

*
[`src/promptflow/promptflow/_sdk/entities/_eager_flow.py`](diffhunk://#diff-53d7ff49a8720f52ee715a6f63d5e389750e44c578a907076d435e541f0163eeR99-R110):
A new method `_init_executable` was added to the `EagerFlow` class to
initialize an `ExecutableEagerFlow`.
*
[`src/promptflow/promptflow/contracts/flow.py`](diffhunk://#diff-b353941bce91518c7f494112d2dd5088b73681eb66e6ff19294edb9c3ec05d0fR887-R929):
The new `EagerFlow` subclass was added to the `FlowBase` class. This
subclass represents an eager flow and includes a static method
`deserialize` to construct an `EagerFlow` from a dictionary.

Changes to FlowInvoker:

*
[`src/promptflow/promptflow/_sdk/_serving/flow_invoker.py`](diffhunk://#diff-8f6d7dd700fff35645eb6488c96b7162914f9a8863d0a38e235b5c9b35cf599dR22-L26):
The `FlowInvoker` class was updated to handle `EagerFlow` and `Flow`
differently during initialization and execution.
[[1]](diffhunk://#diff-8f6d7dd700fff35645eb6488c96b7162914f9a8863d0a38e235b5c9b35cf599dR22-L26)
[[2]](diffhunk://#diff-8f6d7dd700fff35645eb6488c96b7162914f9a8863d0a38e235b5c9b35cf599dL63-R63)
[[3]](diffhunk://#diff-8f6d7dd700fff35645eb6488c96b7162914f9a8863d0a38e235b5c9b35cf599dR77-R79)
[[4]](diffhunk://#diff-8f6d7dd700fff35645eb6488c96b7162914f9a8863d0a38e235b5c9b35cf599dR134-R143)

Changes to Flow entities:

*
[`src/promptflow/promptflow/_sdk/entities/_flow.py`](diffhunk://#diff-339752a6e9c04159a62708efed6b385bd91b016a0f0d2edeb32024fd19886a2bR203-L209):
The `_init_executable` method was updated to initialize an
`ExecutableFlow` differently based on whether the flow is a DAG flow or
not.

Addition of test cases:

*
[`src/promptflow/tests/sdk_cli_test/e2etests/test_flow_serve.py`](diffhunk://#diff-8f175978a83627600049f94831e06a3c9e3307cbde190fe5fbb187571ac514baR356-R365):
A new test case `test_eager_flow_serve` was added to verify the behavior
of the new `EagerFlow`.
*
[`src/promptflow/tests/sdk_cli_test/e2etests/test_flow_test.py`](diffhunk://#diff-acf3c2285ac2a7051d59e3a69abe02224e2d6d5d69e3a0aabd2afd9311f87657R359-R370):
A new test case `test_init_executable` was added to verify the
initialization of an `ExecutableFlow`.

Minor changes:

*
[`src/promptflow/promptflow/_sdk/_serving/flow_invoker.py`](diffhunk://#diff-8f6d7dd700fff35645eb6488c96b7162914f9a8863d0a38e235b5c9b35cf599dR188):
A TODO comment was added to the `invoke` method.
*
[`src/promptflow/promptflow/_sdk/entities/_eager_flow.py`](diffhunk://#diff-53d7ff49a8720f52ee715a6f63d5e389750e44c578a907076d435e541f0163eeR10):
The `generate_flow_meta` function was imported.
*
[`src/promptflow/tests/sdk_cli_test/conftest.py`](diffhunk://#diff-9dced3762e0027d9b6b131576702c5290a82d72c20b2a828c67aa682d31d8e40R41):
The `simple_eager_flow` fixture was added, and the
`create_client_by_model` function was updated to accept a `model_root`
parameter.
[[1]](diffhunk://#diff-9dced3762e0027d9b6b131576702c5290a82d72c20b2a828c67aa682d31d8e40R41)
[[2]](diffhunk://#diff-9dced3762e0027d9b6b131576702c5290a82d72c20b2a828c67aa682d31d8e40L155-R163)
[[3]](diffhunk://#diff-9dced3762e0027d9b6b131576702c5290a82d72c20b2a828c67aa682d31d8e40R213-R217)
*
`src/promptflow/tests/test_configs/eager_flows/simple_with_dict_output/entry.py`
and
`src/promptflow/tests/test_configs/eager_flows/simple_with_dict_output/flow.dag.yaml`:
New test files were added for the `EagerFlow` test case.
[[1]](diffhunk://#diff-a66d84ceff4544e3a46c62757a344589482106f6009f44ef11239f42c4fa2857R1-R7)
[[2]](diffhunk://#diff-5d7098009109f6719af68c34dcf9e1bab9101aec0eed7a5dd5b3f04c50ecafc2R1)

# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.