[promptflow][bugfix] Fix breaking live test by using automatic runtime for resumed run (#2227)
# Description

This PR targets to fix a breaking live test:
<https://github.com/microsoft/promptflow/actions/runs/8162625683/job/22314073969#step:10:2427>

The root cause of this breaking results from the client does not have
permission to access the compute instance that resumed run uses -
automatic runtime is a good way to resolve this, so update the name and
submit it with automatic runtime.

# All Promptflow Contribution checklist:
- [x] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [x] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [x] Title of the pull request is clear and informative.
- [x] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.