[Bugfix] Fix the bug that call flow in flow function will fail (#2267)
# Description

Currently call a flow in a flow node will fail due to wrongly handled
operation context, in this PR, we fix it by copy the original context
and recover.
Issue: https://github.com/microsoft/promptflow/issues/2237

This pull request primarily focuses on refactoring the handling of
operation contexts in the `promptflow` package. The most significant
changes include the addition of a `copy` method to the
`OperationContext` class, the introduction of a `set_instance` class
method to the same class, and changes to the `exec_line_async` and
`_update_operation_context` methods in the `FlowExecutor` class to
utilize these new methods.

Here is a breakdown of the key changes:

Changes to `OperationContext` class in
`src/promptflow/promptflow/_core/operation_context.py`:

* A `copy` method has been added to the `OperationContext` class, which
creates a copy of the current operation context and returns it. This
method also ensures that a copy of the `_OTEL_ATTRIBUTES` is made.
* A `set_instance` class method has been introduced. This method allows
setting a new instance of the operation context.

Changes to `FlowExecutor` class in
`src/promptflow/promptflow/executor/flow_executor.py`:

* In the `exec_line_async` method, the `OperationContext`'s new `copy`
method is used to create a copy of the original operation context.
* The `_update_operation_context` method has been significantly
simplified. Instead of manually reverting changes to the operation
context, it now simply restores the original context using the
`OperationContext`'s `set_instance` method.

These changes simplify the code and make the handling of operation
contexts more robust and less error-prone.

# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.

Co-authored-by: Heyi <heta@microsoft.com>