[Doc] Add environment variable in flow.dag.yaml reference doc (#2263)
# Description

Please add an informative description that covers that changes made by
the pull request and link all relevant issues.
This pull request primarily includes changes to the
`flow-yaml-schema-reference.md` file, which is part of the documentation
for the YAML schema. The changes mainly involve modifications to the
formatting of the tables in the document. The most significant changes
include the adjustment of the column width in several tables and the
addition of a new field, `environment_variables`, in the `nodes`
section.

Formatting changes:

*
[`docs/reference/flow-yaml-schema-reference.md`](diffhunk://#diff-ee48baad7764b577cd9b539588af7e0d9856d7faedd0489dfcf0ec5e3e27bb39L12-R12):
Adjusted the column width in the tables under the `YAML syntax`, `Flow
input`, `Nodes`, and `Node variants` sections. These changes help to
improve the readability of the tables.
[[1]](diffhunk://#diff-ee48baad7764b577cd9b539588af7e0d9856d7faedd0489dfcf0ec5e3e27bb39L12-R12)
[[2]](diffhunk://#diff-ee48baad7764b577cd9b539588af7e0d9856d7faedd0489dfcf0ec5e3e27bb39R21-R28)
[[3]](diffhunk://#diff-ee48baad7764b577cd9b539588af7e0d9856d7faedd0489dfcf0ec5e3e27bb39L47-R48)
[[4]](diffhunk://#diff-ee48baad7764b577cd9b539588af7e0d9856d7faedd0489dfcf0ec5e3e27bb39L64-R65)

Addition of new field:

*
[`docs/reference/flow-yaml-schema-reference.md`](diffhunk://#diff-ee48baad7764b577cd9b539588af7e0d9856d7faedd0489dfcf0ec5e3e27bb39R21-R28):
Added a new field, `environment_variables`, to the `nodes` section. This
field allows users to set environment variables by specifying a property
path and value.
# All Promptflow Contribution checklist:
- [ ] **The pull request does not introduce [breaking changes].**
- [ ] **CHANGELOG is updated for new features, bug fixes or other
significant changes.**
- [ ] **I have read the [contribution guidelines](../CONTRIBUTING.md).**
- [ ] **Create an issue and link to the pull request to get dedicated
review from promptflow team. Learn more: [suggested
workflow](../CONTRIBUTING.md#suggested-workflow).**

## General Guidelines and Best Practices
- [ ] Title of the pull request is clear and informative.
- [ ] There are a small number of commits, each of which have an
informative message. This means that previously merged commits do not
appear in the history of the PR. For more information on cleaning up the
commits in your PR, [see this
page](https://github.com/Azure/azure-powershell/blob/master/documentation/development-docs/cleaning-up-commits.md).

### Testing Guidelines
- [ ] Pull request includes test coverage for the included changes.